Always set `additionalProperties`
This is done to make it easier to distinguish whether additional properties are allowed or not in a schema.

Note: additional properties are always ignored by terraform, and will create a warning if any are set.

External behaviour of the schemas themselves are not changed since `additionalProperties: true` is equivalent to not setting it.