Implement some more CLI
Next up is adding slog and slog options. Also adding unit tests for some of the options.

This PR adds `--output`,`--overwrite`, `--stdout` and `--allow-empty` and implements each of them. There are some edge cases I need to cover with logging and the `stdout` option, but otherwise this is a first draft of the 'complete' CLI.